store_api/users/serializer.py
from rest_framework import serializers
from .models import CustomUser

class CustomUserSerializer(serializers.ModelSerializer):
  
  class Meta:
    model = CustomUser
    fields = ('email', 'first_name', 'last_name', 'is_staff')